migrated refactor tests into lib tests
